requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
SQLAlchemy==2.0.36
asyncpg==0.30.0
psycopg[binary,pool]==3.2.3
pydantic==2.9.2
python-dotenv==1.0.1
httpx==0.27.2
aiogram==3.13.1
apscheduler==3.10.4
pgvector==0.3.6
